backend/requirements.txt
fastapi
uvicorn
asyncpg
asyncio
beautifulsoup4
python-dotenv